BaseEvent class refactored. GlimpseBaseEvent is an interface GlimpseBaseEventAbstract is the abstract class of a generic event that contains just the methods needed to define a generic event GlimpseBaseEventChoreos is an extension of the abstract one with the specification of choreography and service source.
The manager is not yet able to set/read the information related to the source of the messages:
more in deep, all the information that previously was put on the GlimpseBaseEvent like: probeSource, ProbeInstance, eventID(messageID) and all the other information related to the "message" and not to the event.

These parameters must be put on the JMS message when it is created/read by the abstractProbe and by the
ComplexEventProcessor /EventBuffer and so on.

WLT